(refac) Move new process into its own module
