+++ code/index/java/nu/marginalia/index/IndexGrpcService.java
@@ -12,11 +12,11 @@ import nu.marginalia.api.searchquery.*;
 import nu.marginalia.index.model.SearchContext;
 import nu.marginalia.index.model.UnrankedSearchContext;
 import nu.marginalia.index.results.IndexResultRankingService;
-import nu.marginalia.index.searchset.SearchSet;
+import nu.marginalia.ranking.data.SearchSet;
 import nu.marginalia.index.searchset.SearchSetsService;
-import nu.marginalia.index.searchset.SmallSearchSet;
-import nu.marginalia.index.searchset.connectivity.ConnectivitySets;
-import nu.marginalia.index.searchset.connectivity.ConnectivityView;
+import nu.marginalia.ranking.data.SmallSearchSet;
+import nu.marginalia.index.searchset.ConnectivitySets;
+import nu.marginalia.ranking.data.ConnectivityView;
 import nu.marginalia.language.config.LanguageConfiguration;
 import nu.marginalia.language.keywords.KeywordHasher;
 import nu.marginalia.language.model.LanguageDefinition;
