+++ code/index/java/nu/marginalia/index/model/UnrankedSearchContext.java
@@ -15,8 +15,8 @@ import nu.marginalia.api.searchquery.model.results.PrototypeRankingParameters;
 import nu.marginalia.index.CombinedIndexReader;
 import nu.marginalia.index.reverse.IndexLanguageContext;
 import nu.marginalia.index.reverse.query.IndexSearchBudget;
-import nu.marginalia.index.searchset.SearchSet;
-import nu.marginalia.index.searchset.connectivity.ConnectivityView;
+import nu.marginalia.ranking.data.SearchSet;
+import nu.marginalia.ranking.data.ConnectivityView;
 import nu.marginalia.language.keywords.KeywordHasher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
