+++ code/index/test/nu/marginalia/index/IndexQueryServiceIntegrationSmokeTest.java
@@ -13,7 +13,7 @@ import nu.marginalia.index.journal.IndexJournalSlopWriter;
 import nu.marginalia.index.reverse.construction.DocIdRewriter;
 import nu.marginalia.index.reverse.construction.full.FullIndexConstructor;
 import nu.marginalia.index.reverse.construction.prio.PrioIndexConstructor;
-import nu.marginalia.index.searchset.DomainRankings;
+import nu.marginalia.ranking.data.DomainRankings;
 import nu.marginalia.language.keywords.KeywordHasher;
 import nu.marginalia.linkdb.docs.DocumentDbReader;
 import nu.marginalia.linkdb.docs.DocumentDbWriter;
