+++ code/processes/ranking-process/model/java/nu/marginalia/ranking/data/ConnectivityMapFile.java
@@ -1,4 +1,4 @@
-package nu.marginalia.index.searchset.connectivity;
+package nu.marginalia.ranking.data;
 
 import it.unimi.dsi.fastutil.ints.Int2ByteOpenHashMap;
