+++ code/processes/ranking-process/model/java/nu/marginalia/ranking/data/SearchSetAny.java
@@ -1,4 +1,4 @@
-package nu.marginalia.index.searchset;
+package nu.marginalia.ranking.data;
 
 import it.unimi.dsi.fastutil.ints.IntList;
