+++ code/processes/ranking-process/test/nu/marginalia/domainranking/SmallAdjacency.java
@@ -4,7 +4,7 @@ import it.unimi.dsi.fastutil.ints.*;
 import it.unimi.dsi.fastutil.longs.LongOpenHashSet;
 import it.unimi.dsi.fastutil.longs.LongSet;
 import nu.marginalia.array.LongArrayFactory;
-import nu.marginalia.index.searchset.RankingSearchSet;
+import nu.marginalia.ranking.data.RankingSearchSet;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
