+++ code/processes/readme.md
@@ -30,6 +30,7 @@ the data generated by the loader.
 * Ping Process: The [ping-process](ping-process/) keeps track of the aliveness of websites, gathering fingerprint information about the security posture of the website, as well as DNS information.
 * New Domain Process (NDP): The [new-domain-process](new-domain-process/) evaluates new domains for inclusion in the search engine index.
 * Live-Crawling Process: The [live-crawling-process](live-crawling-process/) is a process that crawls websites in real-time based on RSS feeds, updating a smaller index with the latest content.
+* Ranking process: The [ranking-process](ranking-process/) calculates the domain rankings and search sets.
 
 ## Overview
