+++ code/tools/integration-test/test/nu/marginalia/IntegrationTest.java
@@ -20,9 +20,9 @@ import nu.marginalia.index.model.SearchContext;
 import nu.marginalia.index.results.IndexResultRankingService;
 import nu.marginalia.index.reverse.construction.full.FullIndexConstructor;
 import nu.marginalia.index.reverse.construction.prio.PrioIndexConstructor;
-import nu.marginalia.index.searchset.DomainRankings;
-import nu.marginalia.index.searchset.SearchSetAny;
-import nu.marginalia.index.searchset.connectivity.ConnectivityView;
+import nu.marginalia.ranking.data.DomainRankings;
+import nu.marginalia.ranking.data.SearchSetAny;
+import nu.marginalia.ranking.data.ConnectivityView;
 import nu.marginalia.io.SerializableCrawlDataStream;
 import nu.marginalia.language.config.LanguageConfiguration;
 import nu.marginalia.language.keywords.KeywordHasher;
